# twinkle-server/http-api/src/main/resources/openapi/public/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Twinkle Public API v1
  version: 0.1.0
  description: >-
    Stable public v1 contract. Authorization, scope and server-resource
    checks are repeated by the server for every execution. Credential secrets never appear in URLs.
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerCredential: []
paths:
...
  /api/v1/account/{name}:
    get:
      operationId: getAccountByName
      x-required-scope: game:read
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Persisted account summary
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountResponse'}
        '404': {description: Account not found}
  /api/v1/account/{accountId}/characters:
    get:
      operationId: listAccountCharacters
      x-required-scope: game:read
      parameters:
        - {name: accountId, in: path, required: true, schema: {type: integer, format: int64}}
      responses:
        '200':
          description: Characters persisted for the account in the default world
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/CharacterResponse'}
  /api/v1/characters/{characterId}/session:
    delete:
      operationId: disconnectCharacterSession
      x-required-scope: game:write
      parameters:
        - {name: characterId, in: path, required: true, schema: {type: integer, format: int64}}
      responses:
        '204': {description: Character session disconnected}
        '404': {description: Character is not online}
...
components:
  securitySchemes:
    bearerCredential:
      type: http
      scheme: bearer
      bearerFormat: twinkle-credential
...
  schemas:
...
    AccountResponse:
      type: object
      required: [id, name, banned, gender, characterslots]
      properties:
        id: {type: integer, format: int64}
        name: {type: string}
        banned: {type: boolean}
        gender: {type: integer}
        characterslots: {type: integer}
    CharacterResponse:
      type: object
      required: [id, name, level, job, map, meso]
      properties:
        id: {type: integer, format: int64}
        name: {type: string}
        level: {type: integer}
        job: {type: integer}
        map: {type: integer}
        meso: {type: integer}
...